# web/public/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a small set of rest API endpoints using PHP(Lumen) that can be used for listing the names of books along with their authors and comment count, adding and listing anonymous comments for a book, and getting the character list for a book.",
        "version": "1.0.0",
        "title": "Book, Characters & Comments API",
        "termsOfService": "http://landearn.com",
        "contact": {
            "email": "joseph@landearn.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "ec2-13-246-17-45.af-south-1.compute.amazonaws.com",
    "basePath": "/api/v1",
...
    "schemes": ["http"],
    "paths": {
...
        "/comments": {
            "get": {
                "tags": ["comments"],
                "summary": "Get all comments made for a book",
                "description": "Supply the 'book_id' parameter in the url to see all comments made for a book",
                "operationId": "getComments",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "book_id",
                        "in": "query",
                        "description": "Book id",
                        "required": true,
                        "type": "integer",
                        "collectionFormat": "multi"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Comment"
                            }
                        }
                    },
                    "400": {
                        "description": "Please supply the book_id parameter, it cannot be empty or null"
                    },
                    "500": {
                        "description": "A server error occured"
                    }
                }
            },
            "post": {
                "tags": ["comments"],
                "summary": "Post a new comment for a book",
                "description": "Submit a new comment using form data",
                "operationId": "postComment",
                "consumes": ["application/x-www-form-urlencoded"],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "book_id",
                        "in": "formData",
                        "description": "ID of book that owns the comment",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "comment",
                        "in": "formData",
                        "description": "Comment made by anonymous user",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "comment_parent",
                        "in": "formData",
                        "description": "ID of the parent comment for replies under a comment",
                        "required": false,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Comment"
                            }
                        }
                    },
                    "400": {
                        "description": "Book ID and comment cannot be empty"
                    },
                    "500": {
                        "description": "A server error occured"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Comment": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64",
                    "example": 1
                },
                "book_id": {
                    "type": "integer",
                    "format": "int64",
                    "example": 1
                },
                "anonymous_user": {
                    "type": "string",
                    "example": "user_txm9o1qx"
                },
                "comment": {
                    "type": "string",
                    "example": "This is just to confirm that comment works"
                },
                "comment_parent": {
                    "type": "string",
                    "example": null
                },
                "created_at": {
                    "type": "string",
                    "example": "2022-03-30T14:54:18.000000Z"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2022-03-30T14:54:18.000000Z"
                }
            }
        }
    },
...
}